+++ app/src/main/java/com/joaoepedro/tasks/domain/RewardBalanceService.kt
@@ -3,13 +3,13 @@ package com.joaoepedro.tasks.domain
 import com.joaoepedro.tasks.data.RewardTransaction
 
 class RewardBalanceService {
-    fun balanceFor(personId: String, transactions: List<RewardTransaction>): Int {
+    fun balanceFor(personId: String, transactions: List<RewardTransaction>): Double {
         return transactions
             .filter { it.personId == personId }
             .sumOf { it.amount * it.type.sign }
     }
 
-    fun validate(amount: Int, reason: String) {
+    fun validate(amount: Double, reason: String) {
         require(amount > 0) { "O valor precisa ser maior que zero." }
         require(reason.isNotBlank()) { "Informe um motivo." }
     }
